backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
googlemaps
google-api-python-client